+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # GoStorage Changelog
 Все изменения библиотеки GoStorage будут документироваться на этой странице.
 
+## 2024-08-10
+### Added
+- Добавлен параметр `mrpostgres.Options.DSN` для указания соединения с БД в виде URL;
+
 ## 2024-08-03
 ### Added
 - Добавлены комментарии для `mrpostgres.SQLBuilderWhere`;